// client/src/const/msgs.ts
// 消息列表

export enum subMsgs {
  title= '订阅节点',
  addModalText = '添加订阅节点',
  addModalCancelText = '取消',
  addModalOkText = '确定添加',
  addModalNodeURL = '订阅地址(必填)',
  addModalNodeNote = '订阅备注(可不填)',
  addModalNodeFillError = '填写错误,请确认订阅地址是否正确📗',
  addModalTestConnButton = '测试订阅节点连接',
  addNodeModalSubTestStatusSucess = '获取成功',
  addNodeModalSubTestStatusFail = '获取失败',
  getNodeListSuccess = '订阅节点获取成功',
  getNodeListFail = '订阅节点获取失败',
  getNodeDefaultNote = '未命名节点',
  removeAllSubNodeLinkSucess = '删除所有订阅节点成功',
  removeAllSubNodeLinkFail = '删除所有订阅节点失败',
  removeAllSubNodeLinkModalTitle = '删除节点(危险操作)',
  removeAllSubNodeLinkModalContent = '请确定你的操作, 执行此操作将会删除所有的订阅节点',
  removeAllSubNodeLinkModalCancelTitle = '取消删除操作',
  removeAllSubNodeLinkModalCancelContent = '您已经取消删除的操作了',
  subNodesEmptyDesc = '暂无节点',
  subNodesEmptyButton = '更新当前节点'
}